AppointAI-main/requirements.txt
# Google Cloud & AI
google-generativeai>=0.3.0
google-cloud-aiplatform>=1.38.0
google-cloud-vision>=3.4.0
google-cloud-storage>=2.10.0
google-cloud-firestore>=2.13.0

# Web Automation
playwright>=1.40.0
selenium>=4.15.0

# Web Framework
fastapi>=0.104.0
uvicorn>=0.24.0
pydantic>=2.5.0

# Utilities
python-dotenv>=1.0.0
pillow>=10.0.0
requests>=2.31.0
aiohttp>=3.9.0

# Audio/Speech (for voice features)
SpeechRecognition>=3.10.0
pyttsx3>=2.90
pyaudio>=0.2.13

# Data handling
python-multipart>=0.0.6
